feat: provide createWebWorker
So compatible web workers can be created explicitly when desired.
