+++ packages/core/typescript/itk-wasm/src/pipeline/create-worker-proxy.ts
@@ -1,7 +1,7 @@
-import axios from 'axios'
 import * as Comlink from 'comlink'
 
 import WorkerProxy from './web-workers/worker-proxy.js'
+import createWebWorker from './create-web-worker.js'
 
 interface ItkWorker extends Worker {
   workerProxy: WorkerProxy
@@ -25,7 +25,7 @@ function workerToWorkerProxy (worker: Worker): createWorkerProxyResult {
   return { workerProxy, worker: itkWebWorker }
 }
 
-// Internal function to create a web worker promise
+// Internal function to create a web worker proxy
 async function createWorkerProxy (existingWorker: Worker | null, pipelineWorkerUrl?: string | null): Promise<createWorkerProxyResult> {
   let workerProxy: WorkerProxy
   if (existingWorker != null) {
@@ -40,22 +40,7 @@ async function createWorkerProxy (existingWorker: Worker | null, pipelineWorkerU
     }
   }
 
-  const workerUrl = pipelineWorkerUrl
-  let worker = null
-  if (workerUrl === null) {
-    // Use the version built with the bundler
-    //
-    // Bundlers, e.g. WebPack, Vite, Rollup, see these paths at build time
-    worker = new Worker(new URL('./web-workers/itk-wasm-pipeline.worker.js', import.meta.url), { type: 'module' })
-  } else {
-    if ((workerUrl as string).startsWith('http')) {
-      const response = await axios.get((workerUrl as string), { responseType: 'blob' })
-      const workerObjectUrl = URL.createObjectURL(response.data as Blob)
-      worker = new Worker(workerObjectUrl, { type: 'module' })
-    } else {
-      worker = new Worker((workerUrl as string), { type: 'module' })
-    }
-  }
+  const worker = await createWebWorker(pipelineWorkerUrl)
 
   return workerToWorkerProxy(worker)
 }
