+++ packages/core/typescript/itk-wasm/src/pipeline/index.ts
@@ -1,6 +1,7 @@
 // itk-wasm Browser pipeline functions
 
 export { default as runPipeline } from './run-pipeline.js'
+export { default as createWebWorker } from './create-web-worker.js'
 
 export * from './pipeline-worker-url.js'
 export * from './pipelines-base-url.js'
